replace blockNumber with timestamp for canonicalTxChain
